# server/openapi.yaml
openapi: 3.1.0
info:
  title: Stars Trail API
  version: v1
  description: |
    Contract between the Android app (Kotlin) and the Go server (spec §3.2); the two share no code.
    Every path is under /v1. The app sends on every request:
    - `X-Device-Id`: an anonymous per-install ID, the key for the weather proxy's daily quota (falls back to the caller's IP).
    - `X-Client-Version`: its Android versionCode. Below `minClientVersion` every route except
      /v1/health and /v1/version answers 426; offline features never depend on the server.
      A request without the header is not gated (curl, monitoring), so the app must always send it:
      an app that forgets it can never be forced to update.
    Every future feature path lists both headers and the 426 response.
    Routes marked with bearerAuth need `Authorization: Bearer <token>` from /v1/auth/login; a missing,
    unknown or logged-out token is 401 unauthorized. Everything else is anonymous.
    Every error body is an Error: a code from ErrorCode plus the fields that code documents (ADR 0004).
    The Go server is generated from this file (`go generate ./...`, oapi-codegen): edit here first.
servers:
  - url: /v1
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, description: 'the token from /v1/auth/login; valid until logout'}
  parameters:
    DeviceId: {name: X-Device-Id, in: header, schema: {type: string}}
    ClientVersion: {name: X-Client-Version, in: header, schema: {type: integer}}
    TeamId: {name: id, in: path, required: true, schema: {type: integer, format: int64}}
    After: {name: after, in: query, schema: {type: integer, format: int64}, description: 'a Team''s cursor: only positions and messages stored after it'}
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ErrorCode'}
        minClientVersion: {type: integer, description: set when error is client_outdated}
        maxAreaKm2: {type: integer, description: set when error is region_too_large}
        quotaBytes: {type: integer, format: int64, description: set when error is photo_quota_exceeded}
        quotaCells: {type: integer, format: int64, description: set when error is daily_quota_exceeded on weather}
...
    TeamRequest:
      type: object
      description: Nothing to give; members are shown by their account's nickname (a name sent by older apps is ignored).
    JoinRequest:
      type: object
      required: [code]
      properties:
        code: {type: string, pattern: '^[0-9]{4}$'}
    Team:
      type: object
      description: |
        A 队伍 as the caller sees it. Positions and messages are those stored after the `after` cursor the
        request gave (all of them without one); `cursor` is the last stored so far, to pass as `after` next time.
      required: [id, code, initiator, me, ended, cursor, members, messages]
      properties:
        id: {type: integer, format: int64}
        me: {type: integer, format: int64, description: the caller's user id, to find themselves among the members}
        code: {type: string, description: 4-digit join code; free for another team once the trip ends}
        initiator: {type: integer, format: int64, description: user id of the 发起人}
        ended: {type: boolean, description: 结束行程 happened; nobody shares any more}
        cursor: {type: integer, format: int64}
        members: {type: array, items: {$ref: '#/components/schemas/Member'}}
        messages: {type: array, items: {$ref: '#/components/schemas/Message'}, description: '队伍对话, in the order stored'}
        track: {$ref: '#/components/schemas/TeamTrackRef'}
    TeamTrackRef:
      type: object
      description: 'The 队伍轨迹 (spec §2.11) without its points, which GET /teams/{id}/track has; absent when there is none.'
      required: [version, uuid, name, reversed, start]
      properties:
        version: {type: integer, format: int64, description: 'grows with every 指定, 更换, 改起算点 and 取消: a member fetches the points when it changes'}
        uuid: {type: string, maxLength: 64, description: 'the 发起人''s track it was taken from: the same track goes into a member''s 我的轨迹 once'}
        name: {type: string, minLength: 1, maxLength: 100}
        reversed: {type: boolean, description: '起算点: 反向'}
        start: {type: number, format: double, minimum: 0, description: 'a loop''s 起点, metres along the points as they go (0: their own start)'}
...
    Member:
      type: object
      required: [id, name, sharing, positions]
      properties:
        id: {type: integer, format: int64, description: user id}
        name: {type: string, description: their account's nickname as it is now}
        sharing: {type: boolean, description: false after 停止共享 (or 结束行程)}
        positions: {type: array, items: {$ref: '#/components/schemas/Position'}, description: 'in the order stored; 尾迹 sorts by time'}
    Position:
      type: object
      required: [time, lat, lon]
      properties:
        time: {type: integer, format: int64, description: 'fix time, Unix seconds'}
        lat: {type: number, format: double, minimum: -90, maximum: 90}
        lon: {type: number, format: double, minimum: -180, maximum: 180}
        battery: {type: integer, minimum: 0, maximum: 100, description: percent}
    Positions:
      type: object
      required: [positions]
      properties:
        positions: {type: array, minItems: 1, maxItems: 1000, items: {$ref: '#/components/schemas/Position'}}
    MessageKind:
      type: string
      enum:
        - text       # text
        - location   # lat, lon, along
        - image      # image, from /teams/{id}/images
        - system     # text: the server's own note (队伍轨迹 given, changed or dropped), from the member it's about; not sendable
...
    Message:
      type: object
      description: A MessageRequest as stored, with who sent it and when.
      required: [seq, name, time, kind]
      properties:
        seq: {type: integer, format: int64, description: 'order within the team, comparable with a Team''s cursor: unread are those after the last one read'}
        from: {type: integer, format: int64, description: 'sender''s user id; absent once their account is deleted'}
        name: {type: string, description: 'the sender''s nickname as it is now (they may have left since); 已注销用户 once their account is deleted'}
        time: {type: integer, format: int64, description: 'server time, Unix seconds'}
        kind: {$ref: '#/components/schemas/MessageKind'}
        text: {type: string}
        lat: {type: number, format: double}
        lon: {type: number, format: double}
        image: {type: string}
        along: {type: array, items: {type: number, format: double}}
...
    Sharing:
      type: object
      required: [sharing]
      properties:
        sharing: {type: boolean}
...
  responses:
...
    ClientOutdated:
      description: Client older than minClientVersion (client_outdated); the app shows "update required" on this online feature.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Internal:
      description: Server-side failure (internal).
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Unauthorized:
      description: No, unknown or logged-out bearer token (unauthorized); the app logs out locally.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
...
    TeamNotFound:
      description: No such team, or the caller isn't in it (team_not_found).
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
...
paths:
...
  /teams:
    post:
      operationId: postTeam
      summary: Create a 队伍 (spec §2.11); the caller is its 发起人 and gets a 4-digit code to hand out
      description: A caller already in another active team leaves it first. The caller starts out sharing.
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/DeviceId'}, {$ref: '#/components/parameters/ClientVersion'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/TeamRequest'}}}
      responses:
        '200': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Team'}}}}
        '400': {description: invalid_request, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '426': {$ref: '#/components/responses/ClientOutdated'}
        '500': {$ref: '#/components/responses/Internal'}
  /teams/join:
    post:
      operationId: postTeamJoin
      summary: Join the active team with this code (any other active team is left first)
      description: Joining the team one is already in just answers it. A new member starts out sharing.
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/DeviceId'}, {$ref: '#/components/parameters/ClientVersion'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/JoinRequest'}}}
      responses:
        '200': {description: Joined, content: {application/json: {schema: {$ref: '#/components/schemas/Team'}}}}
        '400': {description: invalid_request, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '404': {$ref: '#/components/responses/TeamNotFound'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '426': {$ref: '#/components/responses/ClientOutdated'}
        '500': {$ref: '#/components/responses/Internal'}
  /teams/{id}:
    get:
      operationId: getTeam
      summary: The team, with the positions stored after `after`
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/DeviceId'}, {$ref: '#/components/parameters/ClientVersion'}, {$ref: '#/components/parameters/TeamId'}, {$ref: '#/components/parameters/After'}]
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/Team'}}}}
        '404': {$ref: '#/components/responses/TeamNotFound'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '426': {$ref: '#/components/responses/ClientOutdated'}
        '500': {$ref: '#/components/responses/Internal'}
...
  /teams/{id}/positions:
    post:
      operationId: postTeamPositions
      summary: Report the caller's positions to the team
      description: |
        Live, one at a time; back from no signal, the latest first and then the thinned trail. Positions
        sent while not sharing are dropped.
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/DeviceId'}, {$ref: '#/components/parameters/ClientVersion'}, {$ref: '#/components/parameters/TeamId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/Positions'}}}
      responses:
        '204': {description: Stored and sent to the team}
        '400': {description: invalid_request, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '404': {$ref: '#/components/responses/TeamNotFound'}
        '409': {description: team_ended, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '426': {$ref: '#/components/responses/ClientOutdated'}
        '500': {$ref: '#/components/responses/Internal'}
  /teams/{id}/sharing:
    put:
      operationId: putTeamSharing
      summary: 停止共享 (false) or share again (true); the caller stays in the team either way
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/DeviceId'}, {$ref: '#/components/parameters/ClientVersion'}, {$ref: '#/components/parameters/TeamId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/Sharing'}}}
      responses:
        '204': {description: Set}
        '400': {description: invalid_request, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '404': {$ref: '#/components/responses/TeamNotFound'}
        '409': {description: team_ended, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '426': {$ref: '#/components/responses/ClientOutdated'}
        '500': {$ref: '#/components/responses/Internal'}
  /teams/{id}/leave:
    post:
      operationId: postTeamLeave
      summary: 退出队伍; the caller's 尾迹 goes with them. The last one out ends the trip.
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/DeviceId'}, {$ref: '#/components/parameters/ClientVersion'}, {$ref: '#/components/parameters/TeamId'}]
      responses:
        '204': {description: Left}
        '404': {$ref: '#/components/responses/TeamNotFound'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '426': {$ref: '#/components/responses/ClientOutdated'}
        '500': {$ref: '#/components/responses/Internal'}
  /teams/{id}/end:
    post:
      operationId: postTeamEnd
      summary: 结束行程 (发起人 only); everyone stops sharing, the team stays for its 队伍对话
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/DeviceId'}, {$ref: '#/components/parameters/ClientVersion'}, {$ref: '#/components/parameters/TeamId'}]
      responses:
        '204': {description: Ended (or already was)}
        '403': {description: not_initiator, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '404': {$ref: '#/components/responses/TeamNotFound'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '426': {$ref: '#/components/responses/ClientOutdated'}
        '500': {$ref: '#/components/responses/Internal'}
...